feat(vectors): allow to get gifs

+++ vectors/models.py
@@ -3,6 +3,8 @@ from xml.dom import minidom
 
 from django.core.files.storage import default_storage
 from django.db import models
+from django.db.models import Q
+from django.db.models.constraints import CheckConstraint
 from django.dispatch import receiver
 from django.utils.functional import cached_property
 
@@ -15,21 +17,41 @@ class Vector(models.Model):
     name = models.CharField(max_length=100, blank=True, null=True)
     description = models.CharField(max_length=100, blank=True, null=True)
     tags = TaggableManager()
-    svg = models.FileField()
-    colored_svg = models.FileField(blank=True, null=True)
-    stroke_color = ColorField(blank=True, null=True)
-    fill_color = ColorField(blank=True, null=True)
     uploaded = models.DateTimeField(auto_now_add=True)
     search_text = models.TextField(null=True, blank=True)
 
+    # SVG files
+    svg = models.FileField(blank=True, null=True)
+    colored_svg = models.FileField(blank=True, null=True)
+    stroke_color = ColorField(blank=True, null=True)
+    fill_color = ColorField(blank=True, null=True)
+
+    # GIF files
+    gif = models.FileField(blank=True, null=True)
+    colored_gif = models.FileField(blank=True, null=True)
+
     class Meta:
         ordering = ["id"]
+        constraints = [
+            CheckConstraint(
+                check=(
+                    Q(svg__isnull=False) |
+                    Q(colored_svg__isnull=False) |
+                    Q(gif__isnull=False)  |
+                    Q(colored_gif__isnull=False)
+                ) ,
+                name='vector_should_have_some_file'
+            )
+        ]
 
     def __str__(self):
-        return self.svg.name
+        return self.description
 
     @cached_property
     def svg_content(self):
+        if not self.svg:
+            return ""
+
         with open(self.svg.path, 'r') as f:
             text = f.read()
             return text
@@ -84,6 +106,10 @@ class Featured(models.Model):
         return f"{self.name} ({self.tag or ''})"
 
 
+##################################################################
+# SIGNALS
+##################################################################
+
 @receiver(models.signals.m2m_changed, sender=Vector.tags.through)
 def auto_generate_search_data(sender, instance, action, **kwargs):
     instance.recalculate_search_text()
@@ -91,13 +117,19 @@ def auto_generate_search_data(sender, instance, action, **kwargs):
 
 @receiver(models.signals.post_delete, sender=Vector)
 def auto_delete_svg_on_deletion(sender, instance, **kwargs):
-    if instance.svg:
-        if os.path.isfile(instance.svg.path):
-            default_storage.delete(instance.svg.path)
+    # SVG files
+    if instance.svg and os.path.isfile(instance.svg.path):
+        default_storage.delete(instance.svg.path)
 
-    if instance.colored_svg:
-        if os.path.isfile(instance.colored_svg.path):
-            default_storage.delete(instance.colored_svg.path)
+    if instance.colored_svg and os.path.isfile(instance.colored_svg.path):
+        default_storage.delete(instance.colored_svg.path)
+
+    # GIF files
+    if instance.gif and os.path.isfile(instance.gif.path):
+        default_storage.delete(instance.gif.path)
+
+    if instance.colored_gif and os.path.isfile(instance.colored_gif.path):
+        default_storage.delete(instance.colored_gif.path)
 
 
 @receiver(models.signals.pre_save, sender=Vector)
@@ -110,14 +142,24 @@ def auto_delete_file_on_update(sender, instance, **kwargs):
     except Vector.DoesNotExist:
         return False
 
+    # Svg files
     old_svg = old_vector.svg
     new_svg = instance.svg
-    if old_svg != new_svg:
-        if os.path.isfile(old_svg.path):
-            default_storage.delete(old_svg.path)
+    if old_svg and old_svg != new_svg and os.path.isfile(old_svg.path):
+        default_storage.delete(old_svg.path)
 
     old_colored_svg = old_vector.colored_svg
     new_colored_svg = instance.colored_svg
-    if old_colored_svg and old_colored_svg != new_colored_svg:
-        if os.path.isfile(old_colored_svg.path):
-            default_storage.delete(old_colored_svg.path)
+    if old_colored_svg and old_colored_svg != new_colored_svg and os.path.isfile(old_colored_svg.path):
+        default_storage.delete(old_colored_svg.path)
+
+    # Gif files
+    old_gif = old_vector.gif
+    new_gif = instance.gif
+    if old_gif and old_gif != new_gif and os.path.isfile(old_gif.path):
+        default_storage.delete(old_gif.path)
+
+    old_colored_gif = old_vector.colored_gif
+    new_colored_gif = instance.colored_gif
+    if old_colored_gif and old_colored_gif != new_colored_gif and os.path.isfile(old_colored_gif.path):
+        default_storage.delete(old_colored_gif.path)
